+++ Package.swift
@@ -7,11 +7,11 @@ let package = Package(
     name: "hummingbird",
     platforms: [.macOS(.v14), .iOS(.v17), .tvOS(.v17)],
     products: [
-        .library(name: "Hummingbird", targets: ["Hummingbird"]),
-        .library(name: "HummingbirdFoundation", targets: ["HummingbirdFoundation"]),
-        .library(name: "HummingbirdJobs", targets: ["HummingbirdJobs"]),
-        .library(name: "HummingbirdXCT", targets: ["HummingbirdXCT"]),
-        .executable(name: "PerformanceTest", targets: ["PerformanceTest"]),
+        // .library(name: "Hummingbird", targets: ["Hummingbird"]),
+        // .library(name: "HummingbirdFoundation", targets: ["HummingbirdFoundation"]),
+        // .library(name: "HummingbirdJobs", targets: ["HummingbirdJobs"]),
+        // .library(name: "HummingbirdXCT", targets: ["HummingbirdXCT"]),
+        // .executable(name: "PerformanceTest", targets: ["PerformanceTest"]),
     ],
     dependencies: [
         .package(url: "https://github.com/apple/swift-async-algorithms.git", from: "0.1.0"),
@@ -26,46 +26,37 @@ let package = Package(
         .package(url: "https://github.com/swift-server/swift-service-lifecycle.git", from: "2.0.0"),
     ],
     targets: [
-        .target(name: "Hummingbird", dependencies: [
-            .byName(name: "HummingbirdCore"),
-            .product(name: "ServiceLifecycle", package: "swift-service-lifecycle"),
-            .product(name: "AsyncAlgorithms", package: "swift-async-algorithms"),
-            .product(name: "Logging", package: "swift-log"),
-            .product(name: "Metrics", package: "swift-metrics"),
-            .product(name: "Tracing", package: "swift-distributed-tracing"),
-            .product(name: "NIOCore", package: "swift-nio"),
-            .product(name: "NIOPosix", package: "swift-nio"),
-            .product(name: "NIOHTTP1", package: "swift-nio"),
-        ]),
-        .target(name: "HummingbirdFoundation", dependencies: [
-            .byName(name: "Hummingbird"),
-            .product(name: "NIOCore", package: "swift-nio"),
-            .product(name: "NIOPosix", package: "swift-nio"),
-            .product(name: "NIOFoundationCompat", package: "swift-nio"),
-        ]),
-        .target(name: "HummingbirdJobs", dependencies: [
-            .byName(name: "Hummingbird"),
-            .product(name: "Logging", package: "swift-log"),
-        ]),
-        .target(name: "HummingbirdXCT", dependencies: [
-            .byName(name: "Hummingbird"),
-            .byName(name: "HummingbirdCoreXCT"),
-            .product(name: "NIOCore", package: "swift-nio"),
-            .product(name: "NIOEmbedded", package: "swift-nio"),
-            .product(name: "NIOPosix", package: "swift-nio"),
-            .product(name: "NIOHTTP1", package: "swift-nio"),
-        ]),
+        /*        .target(name: "Hummingbird", dependencies: [
+             .byName(name: "HummingbirdCore"),
+             .product(name: "ServiceLifecycle", package: "swift-service-lifecycle"),
+             .product(name: "AsyncAlgorithms", package: "swift-async-algorithms"),
+             .product(name: "Logging", package: "swift-log"),
+             .product(name: "Metrics", package: "swift-metrics"),
+             .product(name: "Tracing", package: "swift-distributed-tracing"),
+             .product(name: "NIOCore", package: "swift-nio"),
+             .product(name: "NIOPosix", package: "swift-nio"),
+             .product(name: "NIOHTTP1", package: "swift-nio"),
+         ]),
+         .target(name: "HummingbirdFoundation", dependencies: [
+             .byName(name: "Hummingbird"),
+             .product(name: "NIOCore", package: "swift-nio"),
+             .product(name: "NIOPosix", package: "swift-nio"),
+             .product(name: "NIOFoundationCompat", package: "swift-nio"),
+         ]),
+         .target(name: "HummingbirdJobs", dependencies: [
+             .byName(name: "Hummingbird"),
+             .product(name: "Logging", package: "swift-log"),
+         ]),
+         .target(name: "HummingbirdXCT", dependencies: [
+             .byName(name: "Hummingbird"),
+             .byName(name: "HummingbirdCoreXCT"),
+             .product(name: "NIOCore", package: "swift-nio"),
+             .product(name: "NIOEmbedded", package: "swift-nio"),
+             .product(name: "NIOPosix", package: "swift-nio"),
+             .product(name: "NIOHTTP1", package: "swift-nio"),
+         ]),*/
         .target(name: "HummingbirdCore", dependencies: [
-            .product(name: "Logging", package: "swift-log"),
-            .product(name: "NIOCore", package: "swift-nio"),
-            .product(name: "NIOConcurrencyHelpers", package: "swift-nio"),
-            .product(name: "NIOExtras", package: "swift-nio-extras"),
-            .product(name: "NIOHTTP1", package: "swift-nio"),
-            .product(name: "NIOPosix", package: "swift-nio"),
-            .product(name: "NIOTransportServices", package: "swift-nio-transport-services"),
-            .product(name: "ServiceLifecycle", package: "swift-service-lifecycle"),
-        ]),
-        .target(name: "HummingbirdCoreAsync", dependencies: [
+            .product(name: "AsyncAlgorithms", package: "swift-async-algorithms"),
             .product(name: "Logging", package: "swift-log"),
             .product(name: "NIOCore", package: "swift-nio"),
             .product(name: "NIOConcurrencyHelpers", package: "swift-nio"),
@@ -82,36 +73,36 @@ let package = Package(
             .product(name: "NIOPosix", package: "swift-nio"),
             .product(name: "NIOSSL", package: "swift-nio-ssl"),
         ]),
-        .target(name: "HummingbirdHTTP2", dependencies: [
-            .byName(name: "HummingbirdCore"),
-            .product(name: "NIOCore", package: "swift-nio"),
-            .product(name: "NIOHTTP2", package: "swift-nio-http2"),
-            .product(name: "NIOSSL", package: "swift-nio-ssl"),
-        ]),
+        /*        .target(name: "HummingbirdHTTP2", dependencies: [
+                .byName(name: "HummingbirdCore"),
+                .product(name: "NIOCore", package: "swift-nio"),
+                .product(name: "NIOHTTP2", package: "swift-nio-http2"),
+                .product(name: "NIOSSL", package: "swift-nio-ssl"),
+            ]),*/
         .target(name: "HummingbirdTLS", dependencies: [
             .byName(name: "HummingbirdCore"),
             .product(name: "NIOCore", package: "swift-nio"),
             .product(name: "NIOSSL", package: "swift-nio-ssl"),
         ]),
-        .executableTarget(name: "PerformanceTest", dependencies: [
-            .byName(name: "Hummingbird"),
-            .byName(name: "HummingbirdFoundation"),
-            .product(name: "NIOPosix", package: "swift-nio"),
-        ]),
+        /*        .executableTarget(name: "PerformanceTest", dependencies: [
+                .byName(name: "Hummingbird"),
+                .byName(name: "HummingbirdFoundation"),
+                .product(name: "NIOPosix", package: "swift-nio"),
+            ]),*/
         // test targets
-        .testTarget(name: "HummingbirdTests", dependencies: [
-            .byName(name: "Hummingbird"),
-            .byName(name: "HummingbirdFoundation"),
-            .byName(name: "HummingbirdXCT"),
-        ]),
-        .testTarget(name: "HummingbirdFoundationTests", dependencies: [
-            .byName(name: "HummingbirdFoundation"),
-            .byName(name: "HummingbirdXCT"),
-        ]),
-        .testTarget(name: "HummingbirdJobsTests", dependencies: [
-            .byName(name: "HummingbirdJobs"),
-            .byName(name: "HummingbirdXCT"),
-        ]),
+        /*        .testTarget(name: "HummingbirdTests", dependencies: [
+                .byName(name: "Hummingbird"),
+                .byName(name: "HummingbirdFoundation"),
+                .byName(name: "HummingbirdXCT"),
+            ]),
+            .testTarget(name: "HummingbirdFoundationTests", dependencies: [
+                .byName(name: "HummingbirdFoundation"),
+                .byName(name: "HummingbirdXCT"),
+            ]),
+            .testTarget(name: "HummingbirdJobsTests", dependencies: [
+                .byName(name: "HummingbirdJobs"),
+                .byName(name: "HummingbirdXCT"),
+            ]),*/
         .testTarget(
             name: "HummingbirdCoreTests",
             dependencies:
@@ -122,14 +113,6 @@ let package = Package(
                 .product(name: "NIOEmbedded", package: "swift-nio"),
             ],
             resources: [.process("Certificates")]
-        ),
-        .testTarget(
-            name: "HummingbirdCoreAsyncTests",
-            dependencies:
-            [
-                .byName(name: "HummingbirdCoreAsync"),
-                .byName(name: "HummingbirdCoreXCT"),
-            ]
         ),
     ]
 )
