+++ Sources/HummingbirdCoreOld/Error/HTTPError.swift
