+++ Sources/HummingbirdCoreOld/Server/HTTPServer+ServiceLifecycle.swift
